requirements.txt
# Core Agent Framework
smolagents>=1.0.0

# LLM Integration
openai>=1.0.0

# Secure Sandbox Execution
e2b-code-interpreter>=1.0.0

# Web Interface
streamlit>=1.30.0

# Data Processing (for agent computations)
pandas>=2.0.0
numpy>=1.24.0

# Environment Variables
python-dotenv>=1.0.0
